FXML gemerged, Logic und Controller angepasst.
