+++ Catalogue.cfg
@@ -1,0 +1,23 @@
+catalogue {
+  
+	project "Example Project" {
+	 
+	    instruction "Bla";
+	    instruction "Blub";
+	    
+	    test template "SomeTest.java" {
+		    public class SomeTest {
+		    	private String bla = "{\"{";
+		    }
+	    }
+	    
+	    implementation template "SomeClass.java" {
+			public class SomeClass {
+			   //  ... }
+			   /* } */
+			}
+	    }
+	  
+	}
+	  
+}
