--- Main_Controller.java
@@ -1,104 +1,0 @@
-package de.hhu.propra16.project7.controller;
-
-import de.hhu.propra16.project7.catalogue.*;
-
-import java.io.IOException;
-
-import javafx.beans.value.*;
-import javafx.collections.*;
-import javafx.fxml.*;
-import javafx.scene.control.*;
-import javafx.stage.Stage;
-import javafx.event.ActionEvent;
-import javafx.scene.*;
-	 
-public class Main_Controller
-{
-	@FXML
-	private Button startButton, aktButton;
-
-	@FXML
-	private ListView<String> projects;
-
-	@FXML
-	private TextArea goals;
-
-	@FXML
-	private ToggleGroup babyStepsToggleGrp,trackingToggleGrp;
-
-	@FXML
-	private Toggle babyStepsYes,babyStepsNo,trackingYes,trackingNo;
-
-	private ObservableList<String> listViewData = FXCollections.observableArrayList();
-	
-	private Project aktProject;
-
-	private Catalogue catalogue = new Catalogue();
-
-	public Main_Controller(){
-	}
-
-	public void initialize()
-	{
-		fillObservableList();
-
-		// Update aktProject and goals if another Project is clicked
-		projects.getSelectionModel().selectedItemProperty().addListener(new ChangeListener<String>()
-		{
-			
-			public void changed(ObservableValue<? extends String> observable,
-			String oldValue, String newValue)
-			{
-				aktProject = findProject(newValue);
-				goals.setText(aktProject.getInstructions());
-			}
-		});
-	}
-
-	@FXML
-	private void handleAktButtonAction(ActionEvent event) throws IOException
-	{
-		catalogue = new Catalogue();
-		listViewData.clear();
-		fillObservableList();
-	}
-
-	@FXML
-	private void handleStartButtonAction(ActionEvent event) throws IOException
-	{
-		if( aktProject == null ) return;
-		// Wechselt in das nächste Fenster, wenn der Start Button geklickt wird
-		Stage stage = (Stage) startButton.getScene().getWindow();
-		FXMLLoader fxmlloader = new FXMLLoader(getClass().getResource("first.fxml"));
-		if( babyStepsToggleGrp.getSelectedToggle() == babyStepsYes )
-			fxmlloader.setController(new Project_Controller(Status.BabyRed,aktProject));
-		else
-			fxmlloader.setController(new Project_Controller(Status.Red,aktProject));
-		Parent root = fxmlloader.load();
-		Scene scene = new Scene(root, 640, 480);
-		stage.setScene(scene);
-		stage.setResizable(false);
-		stage.setTitle(aktProject.getTitle());
-		stage.show();
-	}
-
-	private Project findProject(String title)
-	{
-		Project project;
-		for( int i = 0 ; i < catalogue.length() ; i++ )
-		{
-			project = catalogue.getProjects().get(i);
-			if( project.getTitle().equals(title) ) return project;
-		}
-		return null;
-	}
-
-	private void fillObservableList()
-	{
-		for( int i = 0 ; i < catalogue.length() ; i++ )
-		{
-			listViewData.add(catalogue.getProjects().get(i).getTitle());
-		}
-		projects.setItems(listViewData);
-	}
-}
