--- Project_Controller.java
@@ -1,97 +1,0 @@
-package de.hhu.propra16.project7.controller;
-
-import de.hhu.propra16.project7.catalogue.*;
-
-import java.io.IOException;
-
-import javafx.fxml.*;
-import javafx.scene.control.*;
-import javafx.stage.Stage;
-import javafx.event.ActionEvent;
-import javafx.scene.*;
-import javafx.scene.text.*;
-import javafx.scene.shape.*;
-import javafx.scene.paint.*;
-	 
-public class Project_Controller
-{
-	@FXML
-	private Button testButton, codeButton, refractoringButton, backButton;
-
-	@FXML
-	private TextArea codewindow;
-
-	@FXML
-	private Circle statusLight;
-
-	@FXML
-	private Text statusAnweisung;
-
-	private Status currStatus;
-	private Project project;
-	private Logic projectLogic = new Logic();
-	
-	public Project_Controller(Status currStatus,Project project)
-	{
-		projectLogic.setStatus(currStatus);
-		this.currStatus = currStatus;
-		this.project = project;
-	}
-
-	public void initialize()
-	{
-		statusLight.setFill(Color.RED);
-		statusAnweisung.setText("Write Tests");
-	}
-	
-	@FXML
-	private void handleTestButtonAction(ActionEvent event) 
-	{
-		projectLogic.Input(Befehl.DoRed);
-		currStatus = projectLogic.getStatus();
-		if( currStatus == Status.Red || currStatus == Status.BabyRed )
-		{
-			codewindow.setText("");
-			statusLight.setFill(Color.RED);
-			statusAnweisung.setText("Write Test");
-		}
-	}
-	
-	@FXML
-	private void handleCodeButtonAction(ActionEvent event) 
-	{
-		projectLogic.Input(Befehl.DoGreen);
-		currStatus = projectLogic.getStatus();
-		if( currStatus == Status.Green || currStatus == Status.BabyGreen )
-		{
-			codewindow.setText("");
-			statusLight.setFill(Color.GREEN);
-			statusAnweisung.setText("Write Code");
-		}
-	}
-	
-	@FXML
-	private void handleRefractoringButtonAction(ActionEvent event)
-	{
-		projectLogic.Input(Befehl.DoRefactoring);
-		currStatus = projectLogic.getStatus();
-		if( currStatus == Status.Refactoring )
-		{
-			statusLight.setFill(Color.BLACK);
-			statusAnweisung.setText("Refactoring");
-		}
-	}
-
-	@FXML
-	private void handleBackButtonAction(ActionEvent event) throws IOException 
-	{
-		// Wechselt in das Main-Fenster, wenn der Start Button geklickt wird
-		Stage stage = (Stage) backButton.getScene().getWindow();
-		Parent root = FXMLLoader.load(getClass().getResource("main_window.fxml"));
-		Scene scene = new Scene(root, 640, 480);
-		stage.setScene(scene);
-		stage.setResizable(false);
-		stage.setTitle("TDDT");
-		stage.show();
-	}
-}
