--- main_window.fxml
@@ -1,86 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-  <?language JavaScript?>
-
-  <?import javafx.scene.control.*?>
-  <?import javafx.scene.layout.*?>
-  <?import javafx.geometry.*?>
-  <?import javafx.scene.text.*?>
-
-  <VBox xmlns:fx="http://javafx.com/fxml" prefWidth="640" prefHeight="480">
-
-    <HBox prefWidth="640" minHeight="10" prefHeight="10">
-    </HBox>
-
-    <HBox prefWidth="640" prefHeight="460">
-
-      <VBox minWidth="20" prefWidth="20" prefHeight="460">
-      </VBox>
-
-      <VBox spacing="10" prefWidth="260" prefHeight="460">
-        <children>
-          <VBox prefWidth="260" prefHeight="20" alignment="CENTER">
-            <Text text="Übungen" textAlignment="CENTER">
-              <font>
-                <Font name="Arial Black" size="20"/>
-              </font>
-            </Text>
-          </VBox>
-          <VBox prefWidth="260" prefHeight="420" alignment="CENTER">
-            <ListView fx:id="projects" prefWidth="260" prefHeight="420">
-            </ListView>
-          </VBox>
-        </children>
-      </VBox>
-
-      <VBox prefWidth="40" prefHeight="460">
-      </VBox>
-
-      <VBox spacing="10" prefWidth="300" prefHeight="460">
-        <children>
-          <VBox prefWidth="300" prefHeight="20" alignment="CENTER">
-            <Text text="Instructions" textAlignment="CENTER">
-              <font>
-                <Font name="Arial Black" size="20"/>
-              </font>
-            </Text>
-          </VBox>
-          <VBox prefWidth="300" prefHeight="300">
-            <TextArea fx:id="goals" editable="false" prefWidth="300" prefHeight="300">
-            </TextArea>
-          </VBox>
-          <VBox prefWidth="300" prefHeight="120">
-            <GridPane alignment="center" hgap="10" vgap="10">
-              <padding>
-                <Insets top="25" right="25" bottom="25" left="25"/>
-              </padding>
-              <Label text="Babysteps: " GridPane.columnIndex="0" GridPane.rowIndex="0"/>
-              <fx:define>
-                <ToggleGroup fx:id="babyStepsToggleGrp"/>
-              </fx:define>
-              <RadioButton fx:id="babyStepsYes" text="Yes" toggleGroup="$babyStepsToggleGrp" GridPane.columnIndex="1" GridPane.rowIndex="0"/>
-              <RadioButton fx:id="babyStepsNo" text="No" toggleGroup="$babyStepsToggleGrp" GridPane.columnIndex="2" GridPane.rowIndex="0"/>
-              <Label text="Tracking: " GridPane.columnIndex="0" GridPane.rowIndex="1"/>
-              <fx:define>
-                <ToggleGroup fx:id="trackingToggleGrp"/>
-              </fx:define>
-              <RadioButton fx:id="trackingYes" text="Yes" toggleGroup="$trackingToggleGrp" GridPane.columnIndex="1" GridPane.rowIndex="1"/>
-              <RadioButton fx:id="trackingNo" text="No" toggleGroup="$trackingToggleGrp" GridPane.columnIndex="2" GridPane.rowIndex="1"/>
-            </GridPane>
-          </VBox>
-        </children>
-      </VBox>
-
-
-      <VBox minWidth="20" prefWidth="20" prefHeight="460">
-      </VBox>
-
-    </HBox>
-
-    <VBox prefWidth="640" prefHeight="80">
-      <HBox spacing="20" prefHeight="80" alignment="CENTER">
-        <Button fx:id="startButton" mnemonicParsing="false" onAction="#handleAktButtonAction" text="Aktualisieren" alignment="CENTER" textAlignment="CENTER"/>
-        <Button fx:id="startButton" mnemonicParsing="false" onAction="#handleStartButtonAction" text="Start" alignment="CENTER" textAlignment="CENTER"/>
-      </HBox>
-    </VBox>
-
-  </VBox>
