--- project_window.fxml
@@ -1,53 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?language JavaScript?>
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.TextArea?>
-<?import javafx.scene.layout.HBox?>
-<?import javafx.scene.layout.TilePane?>
-<?import javafx.scene.layout.VBox?>
-<?import javafx.scene.shape.Circle?>
-<?import javafx.scene.text.Font?>
-<?import javafx.scene.text.Text?>
-
-<HBox prefHeight="480" prefWidth="640" xmlns="http://javafx.com/javafx/8.0.65" xmlns:fx="http://javafx.com/fxml/1" fx:controller="de.hhu.propra16.project7.controller.Project_Controller">
-
-    <VBox prefHeight="480" prefWidth="40" />
-
-    <VBox prefHeight="480" prefWidth="560">
-
-      <VBox prefHeight="10" prefWidth="560" />
-
-      <TilePane alignment="CENTER" prefHeight="40" prefWidth="100">
-        <Text fx:id="statusAnweisung" textAlignment="CENTER">
-          <font>
-            <Font name="Arial Black" size="20.0" />
-          </font>
-        </Text>
-      </TilePane>
-      <VBox prefHeight="350" prefWidth="560">
-        <TextArea fx:id="codewindow" prefHeight="350" prefWidth="560" />
-      </VBox>
-
-      <HBox prefHeight="80" prefWidth="640">
-        <HBox prefHeight="80.0" prefWidth="192.0" />
-        <HBox alignment="CENTER" prefHeight="80" prefWidth="733.0" spacing="20">
-          <Button fx:id="resetButton" alignment="CENTER" mnemonicParsing="false" onAction="#handleTestButtonAction" prefWidth="100" text="Test" textAlignment="CENTER" />
-          <Button fx:id="backButton" alignment="CENTER" mnemonicParsing="false" onAction="#handleCodeButtonAction" prefWidth="100" text="Code" textAlignment="CENTER" />
-          <Button fx:id="weiterButton" alignment="CENTER" mnemonicParsing="false" onAction="#handleRefractoringButtonAction" prefWidth="100" text="Refractoring" textAlignment="CENTER" />
-            <Button alignment="CENTER" mnemonicParsing="false" onAction="#handleBackButtonAction" prefWidth="100" text="Zurück" textAlignment="CENTER" />
-        </HBox>
-      </HBox>
-
-    </VBox>
-
-    <VBox prefHeight="480" prefWidth="40">
-      <VBox prefHeight="430" prefWidth="40" />
-      <TilePane alignment="CENTER" prefWidth="20">
-        <children>
-          <Circle fx:id="statusLight" centerX="10" centerY="10" radius="10" />
-        </children>
-      </TilePane>
-    </VBox>
-
-  </HBox>
