+++ src/main/java/de/hhu/propra16/project7/controller/main_window.fxml
@@ -6,7 +6,7 @@
   <?import javafx.geometry.*?>
   <?import javafx.scene.text.*?>
 
-  <VBox fx:controller="de.hhu.propra16.project7.controller.Main_Controller" xmlns:fx="http://javafx.com/fxml" prefWidth="640" prefHeight="480">
+  <VBox xmlns:fx="http://javafx.com/fxml" prefWidth="640" prefHeight="480" fx:controller="de.hhu.propra16.project7.controller.Main_Controller">
 
     <HBox prefWidth="640" minHeight="10" prefHeight="10">
     </HBox>
@@ -38,18 +38,38 @@
       <VBox spacing="10" prefWidth="300" prefHeight="460">
         <children>
           <VBox prefWidth="300" prefHeight="20" alignment="CENTER">
-            <Text text="Ziele der Übung" textAlignment="CENTER">
+            <Text text="Instructions" textAlignment="CENTER">
               <font>
                 <Font name="Arial Black" size="20"/>
               </font>
             </Text>
           </VBox>
-          <VBox>
-            <TextArea fx:id="goals" prefWidth="300" prefHeight="420">
+          <VBox prefWidth="300" prefHeight="300">
+            <TextArea fx:id="goals" editable="false" prefWidth="300" prefHeight="300">
             </TextArea>
+          </VBox>
+          <VBox prefWidth="300" prefHeight="120">
+            <GridPane alignment="center" hgap="10" vgap="10">
+              <padding>
+                <Insets top="25" right="25" bottom="25" left="25"/>
+              </padding>
+              <Label text="Babysteps: " GridPane.columnIndex="0" GridPane.rowIndex="0"/>
+              <fx:define>
+                <ToggleGroup fx:id="babyStepsToggleGrp"/>
+              </fx:define>
+              <RadioButton fx:id="babyStepsYes" text="Yes" toggleGroup="$babyStepsToggleGrp" GridPane.columnIndex="1" GridPane.rowIndex="0"/>
+              <RadioButton fx:id="babyStepsNo" text="No" toggleGroup="$babyStepsToggleGrp" GridPane.columnIndex="2" GridPane.rowIndex="0"/>
+              <Label text="Tracking: " GridPane.columnIndex="0" GridPane.rowIndex="1"/>
+              <fx:define>
+                <ToggleGroup fx:id="trackingToggleGrp"/>
+              </fx:define>
+              <RadioButton fx:id="trackingYes" text="Yes" toggleGroup="$trackingToggleGrp" GridPane.columnIndex="1" GridPane.rowIndex="1"/>
+              <RadioButton fx:id="trackingNo" text="No" toggleGroup="$trackingToggleGrp" GridPane.columnIndex="2" GridPane.rowIndex="1"/>
+            </GridPane>
           </VBox>
         </children>
       </VBox>
+
 
       <VBox minWidth="20" prefWidth="20" prefHeight="460">
       </VBox>
